requirement.txt
jupyter
pandas
cassandra-driver
